Implemented pagination and ordering search for users, implemented ordering for getAll methods

+++ src/main/java/com/animalmanagementsystem/shelter/services/impl/AnimalServiceImpl.java
@@ -14,6 +14,7 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import jakarta.transaction.Transactional;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -26,6 +27,7 @@ public class AnimalServiceImpl implements AnimalService {
     private final AnimalRepository animalRepository;
     private final AnimalMapper animalMapper;
     private final EntityManager entityManager;
+    public static final String NAME = "name";
 
     public AnimalServiceImpl(AnimalRepository animalRepository, AnimalMapper animalMapper, EntityManager entityManager) {
         this.animalRepository = animalRepository;
@@ -42,13 +44,16 @@ public class AnimalServiceImpl implements AnimalService {
 
         if (request.query() != null && !request.query().isBlank()) {
             String query = "%" + request.query() + "%";
-            Predicate namePredicate = criteriaBuilder.like(root.get("name"), query);
+            Predicate namePredicate = criteriaBuilder.like(root.get(NAME), query);
             Predicate speciesPredicate = criteriaBuilder.like(root.get("species"), query);
 
             predicates.add(criteriaBuilder.or(namePredicate, speciesPredicate));
         }
 
         criteriaQuery.where(criteriaBuilder.or(predicates.toArray(new Predicate[0])));
+
+        criteriaQuery.orderBy(criteriaBuilder.asc(root.get(NAME)));
+
         TypedQuery<AnimalEntity> query = entityManager.createQuery(criteriaQuery);
 
         return query.getResultList().stream().map(animalMapper::mapEntityToDto).toList();
@@ -69,7 +74,7 @@ public class AnimalServiceImpl implements AnimalService {
 
     @Override
     public List<AnimalDto> getAllAnimals() {
-        List<AnimalEntity> animalEntities = animalRepository.findAll();
+        List<AnimalEntity> animalEntities = animalRepository.findAll(Sort.by(Sort.Direction.ASC, NAME));
         return animalEntities.stream().map(animalMapper::mapEntityToDto).toList();
     }
 
@@ -95,7 +100,9 @@ public class AnimalServiceImpl implements AnimalService {
     public void deleteAnimal(Long id) {
         Optional<AnimalEntity> optionalAnimalEntity = animalRepository.findById(id);
 
-        optionalAnimalEntity.orElseThrow(() -> new AnimalNotFoundException(id));
+        if (optionalAnimalEntity.isEmpty()) {
+            throw new AnimalNotFoundException(id);
+        }
 
         animalRepository.deleteById(id);
     }
