+++ src/main/java/com/animalmanagementsystem/shelter/services/impl/HealthServiceImpl.java
@@ -14,6 +14,7 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import jakarta.transaction.Transactional;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.sql.Date;
@@ -28,6 +29,7 @@ public class HealthServiceImpl implements HealthService {
     private final HealthRepository healthRepository;
     private final HealthMapper healthMapper;
     private final EntityManager entityManager;
+    public static final String STATUS = "status";
 
     public HealthServiceImpl(HealthRepository healthRepository, HealthMapper healthMapper, EntityManager entityManager) {
         this.healthRepository = healthRepository;
@@ -43,11 +45,13 @@ public class HealthServiceImpl implements HealthService {
         Root<HealthEntity> healthEntityRoot = criteriaQuery.from(HealthEntity.class);
 
         if (request.status() != null && !request.status().isBlank()) {
-            Predicate statusPredicate = criteriaBuilder.like(healthEntityRoot.get("status"), "%" + request.status() + "%");
+            Predicate statusPredicate = criteriaBuilder.like(healthEntityRoot.get(STATUS), "%" + request.status() + "%");
             predicates.add(statusPredicate);
         }
 
         criteriaQuery.where(criteriaBuilder.and(predicates.toArray(new Predicate[0])));
+
+        criteriaQuery.orderBy(criteriaBuilder.asc(healthEntityRoot.get(STATUS)));
 
         TypedQuery<HealthEntity> query = entityManager.createQuery(criteriaQuery);
 
@@ -70,7 +74,7 @@ public class HealthServiceImpl implements HealthService {
 
     @Override
     public List<HealthDto> getAllHealths() {
-        List<HealthEntity> healthEntities = healthRepository.findAll();
+        List<HealthEntity> healthEntities = healthRepository.findAll(Sort.by(Sort.Direction.ASC, STATUS));
         return healthEntities.stream().map(healthMapper::mapEntityToDto).toList();
     }
 
@@ -93,7 +97,9 @@ public class HealthServiceImpl implements HealthService {
     public void deleteHealth(Long id) {
         Optional<HealthEntity> optionalHealthEntity = healthRepository.findById(id);
 
-        optionalHealthEntity.orElseThrow(() -> new HealthNotFoundException(id));
+        if (optionalHealthEntity.isEmpty()) {
+            throw new HealthNotFoundException(id);
+        }
 
         healthRepository.deleteById(id);
     }
