+++ src/main/java/com/animalmanagementsystem/shelter/services/impl/RoleServiceImpl.java
@@ -14,6 +14,7 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import jakarta.transaction.Transactional;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -26,6 +27,7 @@ public class RoleServiceImpl implements RoleService {
     private final RoleRepository roleRepository;
     private final RoleMapper roleMapper;
     private final EntityManager entityManager;
+    public static final String NAME = "name";
 
     public RoleServiceImpl(RoleRepository roleRepository, RoleMapper roleMapper, EntityManager entityManager) {
         this.roleRepository = roleRepository;
@@ -42,13 +44,15 @@ public class RoleServiceImpl implements RoleService {
 
         if (request.query() != null && !request.query().isBlank()) {
             String query = "%" + request.query() + "%";
-            Predicate namePredicate = criteriaBuilder.like(roleEntityRoot.get("name"), query);
+            Predicate namePredicate = criteriaBuilder.like(roleEntityRoot.get(NAME), query);
             Predicate descriptionPredicate = criteriaBuilder.like(roleEntityRoot.get("description"), query);
 
             predicates.add(criteriaBuilder.or(namePredicate, descriptionPredicate));
         }
 
         criteriaQuery.where(predicates.toArray(new Predicate[0]));
+
+        criteriaQuery.orderBy(criteriaBuilder.asc(roleEntityRoot.get(NAME)));
 
         TypedQuery<RoleEntity> query = entityManager.createQuery(criteriaQuery);
 
@@ -70,7 +74,7 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public List<RoleDto> getAllRoles() {
-        List<RoleEntity> roleEntities = roleRepository.findAll();
+        List<RoleEntity> roleEntities = roleRepository.findAll(Sort.by(Sort.Direction.ASC, NAME));
         return roleEntities.stream().map(roleMapper::mapEntityToDto).toList();
     }
 
@@ -93,7 +97,9 @@ public class RoleServiceImpl implements RoleService {
     public void deleteRole(Long id) {
         Optional<RoleEntity> optionalRoleEntity = roleRepository.findById(id);
 
-        optionalRoleEntity.orElseThrow(() -> new RoleNotFoundException(id));
+        if (optionalRoleEntity.isEmpty()) {
+            throw new RoleNotFoundException(id);
+        }
 
         roleRepository.deleteById(id);
     }
